Rename infoKey to infoSlug in ToolMeta and move AddStepDialog to design route
Actions:
+ rename infoKey property to infoSlug in ToolMeta type definition
+ update all infoKey references to infoSlug across TOOL_META entries
+ update toolInfoUrl function to use meta.infoSlug
+ move AddStepDialog.svelte from lib/components to routes/(admin)/admin/conversations/[conversation_id]/design
+ update AddStepDialog import path in design/+layout.svelte
